Split commands according to public api support
